Added conditional to input routine. Significant performance boost
Previously, python would reject a large amount of data from the
arduino because it was recieving multiple measurements at once.
This forces the input routine to exit as soon as it captures a
single datapoint input (terminated by '\n').

This has increased the sample rate on my laptop from about 175sps
to 350sps