Connected FPS and bold-line menu options to methods
restart_graph_timer has been created to update the timer for
the FPS menu options. The millisecond delay is recalculated,
the timer is stopped, and then started again. It will not
start the timer if the timer isn't currently running.

QActionGroup had to be added manually in the .ui file.
Qt Creator doesn't have a way to group QActions into a
QActionGroup for some reason.